refactor: move config options type to config file
